agents/taskmanager_agent/pyproject.toml
[project]
name = "taskmanager-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.11",
    "click>=8.2.1",
    "dotenv>=0.9.9",
    "logging>=0.4.9.6",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]
